chore: merging serve code
