+++ rust/numaflow-core/src/pipeline.rs
@@ -20,9 +20,9 @@ use crate::pipeline::isb::jetstream::writer::JetstreamWriter;
 use crate::pipeline::pipeline::isb::BufferReaderConfig;
 use crate::shared::create_components;
 use crate::shared::metrics::start_metrics_server;
-use crate::sink::serving_store::ServingStore;
-use crate::sink::serving_store::nats::NatsServingStore;
-use crate::sink::serving_store::user_defined::UserDefinedStore;
+use crate::sink::serve::ServingStore;
+use crate::sink::serve::nats::NatsServingStore;
+use crate::sink::serve::user_defined::UserDefinedStore;
 use crate::tracker::TrackerHandle;
 use crate::watermark::WatermarkHandle;
 use crate::watermark::isb::ISBWatermarkHandle;
