+++ rust/numaflow-core/src/sink/serve/nats.rs
@@ -1,5 +1,5 @@
 use crate::config::pipeline::NatsStoreConfig;
-use crate::sink::serving_store::StoreEntry;
+use crate::sink::serve::StoreEntry;
 use async_nats::jetstream::Context;
 use async_nats::jetstream::kv::Store;
 use chrono::Utc;
